feat: #3412 - keep shadow folder of failed job if --keep-incomplete flag is set. (#3430)
### Description

resolves #3412 

Shadow folders are now kept for failed jobs when the flag
--keep-incomplete is set.
A warning is displayed in the logs with the path to the shadow
directory.

+++ src/snakemake/dag.py
@@ -813,8 +813,9 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
                         rule=job.rule,
                     )
 
-    def unshadow_output(self, job, only_log=False):
+    def unshadow_output(self, job, only_log=False, keep_shadow_dir=False):
         """Move files from shadow directory to real output paths."""
+        """If shadow directory is kept, returns the path of it."""
         if not job.shadow_dir or not job.output:
             return
 
@@ -833,7 +834,11 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
             if os.path.realpath(shadow_output) == os.path.realpath(real_output):
                 continue
             shutil.move(shadow_output, real_output)
-        shutil.rmtree(job.shadow_dir)
+        if keep_shadow_dir and os.listdir(job.shadow_dir):
+            return str(job.shadow_dir)
+        else:
+            shutil.rmtree(job.shadow_dir)
+            return None
 
     def check_periodic_wildcards(self, job):
         """Raise an exception if a wildcard of the given job appears to be periodic,
